+++ docs/source/lumin.nn.models.blocks.rst
@@ -28,10 +28,10 @@ lumin.nn.models.blocks.endcap module
    :undoc-members:
    :show-inheritance:
 
-lumin.nn.models.blocks.gnn\_block module
-----------------------------------------
+lumin.nn.models.blocks.gnn\_blocks module
+-----------------------------------------
 
-.. automodule:: lumin.nn.models.blocks.gnn_block
+.. automodule:: lumin.nn.models.blocks.gnn_blocks
    :members:
    :undoc-members:
    :show-inheritance:
