+++ docs/source/lumin.nn.models.rst
@@ -5,8 +5,7 @@ Subpackages
 -----------
 
 .. toctree::
-   :maxdepth: 1
-   
+
    lumin.nn.models.blocks
    lumin.nn.models.layers
