+++ docs/source/lumin.nn.rst
@@ -5,8 +5,7 @@ Subpackages
 -----------
 
 .. toctree::
-   :maxdepth: 1
-   
+
    lumin.nn.callbacks
    lumin.nn.data
    lumin.nn.ensemble
