+++ docs/source/modules.rst
@@ -2,6 +2,6 @@ lumin
 =====
 
 .. toctree::
-   :maxdepth: 1
+   :maxdepth: 4
 
    lumin
